Move IMSIC out of the PrivilegePlugin
